requirements.txt
anyio==4.9.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.2
cssselect2==0.8.0
exceptiongroup==1.3.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
flet==0.28.3
fonttools==4.58.0
googlesearch-python==1.3.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
nltk==3.9.1
numpy==2.2.6
oauthlib==3.2.2
pandas==2.2.3
pdfminer.six==20250506
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
PyPDF2==3.0.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
repath==0.9.0
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
wikipedia==1.4.0
Wikipedia-API==0.8.1
zopfli==0.2.3.post1
gunicorn==20.1.0